:recycle: refactor: modify authlayout structure
