+++ src/app/(user-access)/login/loginPage.module.css
@@ -19,6 +19,7 @@
 }
 
 .input {
+  width: 100%;
   padding: 8px;
   border: 1px solid var(--gray-300);
   border-radius: 8px;
@@ -38,28 +39,13 @@
   margin-top: 10px;
 }
 
-.logoContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 16px 0;
-  margin-top: 209px;
-}
-
-.logo {
-  width: 50%;
-  height: auto;
-}
-
-.greeting {
-  margin-top: 16px;
-  font-size: 20px;
-  text-align: center;
-}
-
 .disabled {
   background-color: var(--gray-400);
   cursor: not-allowed;
   opacity: 0.5;
+}
+
+.greeting {
+  font-size: 20px;
+  text-align: center;
 }
