+++ src/app/(user-access)/login/page.tsx
@@ -7,8 +7,8 @@ import useAuthStore from '@/store/authStore';
 import Button from '@/components/Button';
 import styles from './loginPage.module.css';
 import axiosInstance from '@/lib/axiosInstance';
-import Image from 'next/image';
 import { ERROR_MESSAGES } from '@/constants/message';
+import AuthLayout from '@/app/(user-access)/AuthLayout';
 
 type LoginFormInputs = {
   email: string;
@@ -48,84 +48,74 @@ export default function LoginPage() {
   };
 
   return (
-    <form onSubmit={handleSubmit(onSubmit)} className={styles.form}>
-      <div className={styles.inputWrapper}>
-        <div className={styles.logoContainer}>
-          <Image
-            src="/images/logo_main.svg"
-            alt="로고"
-            width={100}
-            height={100}
-            className={styles.logo}
-            onClick={() => router.push('/')}
-            style={{ cursor: 'pointer' }}
-          />
+    <AuthLayout>
+      <form onSubmit={handleSubmit(onSubmit)} className={styles.form}>
+        <div className={styles.inputWrapper}>
           <p className={styles.greeting}>오늘도 만나서 반가워요!</p>
+          <label htmlFor="email" className={styles.label}>
+            이메일
+          </label>
+          <input
+            id="email"
+            type="email"
+            className={`${styles.input} ${errors.email ? styles.inputError : ''}`}
+            {...register('email', {
+              required: ERROR_MESSAGES.REQUIRED_EMAIL,
+              pattern: {
+                value: /^[^\s@]+@[^\s@]+\.[^\s@]+$/,
+                message: ERROR_MESSAGES.INVALID_EMAIL,
+              },
+            })}
+          />
+          {errors.email && (
+            <span className={styles.errorMessage}>{errors.email.message}</span>
+          )}
         </div>
 
-        <label htmlFor="email" className={styles.label}>
-          이메일
-        </label>
-        <input
-          id="email"
-          type="email"
-          className={`${styles.input} ${errors.email ? styles.inputError : ''}`}
-          {...register('email', {
-            required: ERROR_MESSAGES.REQUIRED_EMAIL,
-            pattern: {
-              value: /^[^\s@]+@[^\s@]+\.[^\s@]+$/,
-              message: ERROR_MESSAGES.INVALID_EMAIL,
-            },
-          })}
-        />
-        {errors.email && (
-          <span className={styles.errorMessage}>{errors.email.message}</span>
-        )}
-      </div>
-
-      <div className={styles.inputWrapper}>
-        <label htmlFor="password" className={styles.label}>
-          비밀번호
-        </label>
-        <input
-          id="password"
-          type="password"
-          className={`${styles.input} ${errors.password ? styles.inputError : ''}`}
-          {...register('password', {
-            required: ERROR_MESSAGES.PASSWORD_REQUIRE,
-            minLength: {
-              value: 8,
-              message: ERROR_MESSAGES.PASSWORD_TOO_SHORT,
-            },
-          })}
-        />
-        {errors.password && (
-          <p className={styles.errorMessage}>{errors.password.message}</p>
-        )}
-      </div>
-      <Button
-        type="submit"
-        disabled={!isValid}
-        className={!isValid ? styles.disabled : ''}
-        style={{ height: '40px' }}
-      >
-        로그인
-      </Button>
-      <div style={{ display: 'flex', justifyContent: 'center' }}>
-        <p>
-          회원이 아니신가요?{' '}
-          <span
-            style={{
-              color: 'var(--violet)',
-              textDecoration: 'underline',
-              cursor: 'pointer',
-            }}
-            onClick={() => router.push('/signup')}
-          >
-            회원가입하기
-          </span>{' '}
-        </p>
-      </div>
-    </form>
+        <div className={styles.inputWrapper}>
+          <label htmlFor="password" className={styles.label}>
+            비밀번호
+          </label>
+          <input
+            id="password"
+            type="password"
+            className={`${styles.input} ${errors.password ? styles.inputError : ''}`}
+            {...register('password', {
+              required: ERROR_MESSAGES.PASSWORD_REQUIRE,
+              minLength: {
+                value: 8,
+                message: ERROR_MESSAGES.PASSWORD_TOO_SHORT,
+              },
+            })}
+          />
+          {errors.password && (
+            <p className={styles.errorMessage}>{errors.password.message}</p>
+          )}
+        </div>
+        <Button
+          type="submit"
+          disabled={!isValid}
+          className={!isValid ? styles.disabled : ''}
+          style={{ height: '40px' }}
+        >
+          로그인
+        </Button>
+        <div style={{ display: 'flex', justifyContent: 'center' }}>
+          <p>
+            회원이 아니신가요?{' '}
+            <span
+              style={{
+                color: 'var(--violet)',
+                textDecoration: 'underline',
+                cursor: 'pointer',
+              }}
+              onClick={() => router.push('/signup')}
+            >
+              회원가입하기
+            </span>{' '}
+          </p>
+        </div>
+      </form>
+    </AuthLayout>
   );
 }
