--- src/components/AuthLayout.module.css
@@ -1,28 +1,0 @@
-.authContainer {
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-  min-height: 100vh;
-  width: 100%;
-  background-color: var(--gray-100);
-}
-
-.authContent {
-  width: 100%;
-  max-width: 400px;
-  background-color: var(--white);
-  border-radius: 8px;
-  padding: 20px;
-}
-
-@media screen and (min-width: 768px) {
-  .authContent {
-    max-width: 351px;
-  }
-}
-
-@media screen and (min-width: 1200px) {
-  .authContent {
-    max-width: 520px;
-  }
-}
